requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-genai
python-dotenv==0.19.0